refactor: extract ConeMemoryStore, ScoopCostTracker, ScoopCompletionService (refs #1091)
Continues the Orchestrator god-class refactor by carving three more
single-responsibility collaborators off the class:

- ConeMemoryStore (cone-memory-store.ts) owns the /shared/CLAUDE.md +
  /workspace/CLAUDE.md read/write surface and the one-shot legacy
  auto-extracted-block migration. Orchestrator now keeps thin
  getGlobalMemory / setGlobalMemory / appendConeMemory delegators for
  the public API.

- ScoopCostTracker (scoop-cost-tracker.ts) owns per-session cost
  aggregation and the droppedScoopCosts buffer that preserves cost data
  for scoops that have been dropped this session. buildScoopCost is now a
  pure free function.

- ScoopCompletionService (scoop-completion-service.ts) owns the per-scoop
  response buffer, the VFS artifact persistence + scoop-notify routing,
  and the scoop_mute / scoop_wait coordination state (mutedScoops,
  pendingCompletions, completionWaiters). The orchestrator's
  unregisterScoop / handleScoopFatalError / shutdown cleanup paths now
  delegate to forgetScoop(reason) and shutdown() instead of poking four
  maps each.

Behavior is preserved end-to-end: 92 orchestrator tests still pass against
the new module boundaries. Test fixtures that reached into the
orchestrator's private state now navigate through completionService and
memoryStore.

orchestrator.ts is 2129 lines, down from ~3099 before this work.
